requirements.txt
numpy
pandas
Pillow
matplotlib
faiss-cpu
scikit-learn
